Server shows the IP
The server now shows the IP when you start up the connection

Made it so when a client attempts to connect it fails before attempting to write a username.

Added visual looks for cmd. Previously cmd did not support ANSI so I found a very simple way around it

To do:
Possibly add a GUI

Notes:
Program could techincally be finished here. Core functionality added but I want to add more features.

Been working on this since around 5 and it is now 12:33 am. I am going to sleep and add some extra stuff tommorow. Goodnight future me and anybody else that may read this